float

+++ dialogue_manager.py
@@ -34,10 +34,12 @@ def run_an_episode():
 		slot,intent = nlu.predict(nl_input,system)
 		system.update(slot,intent,nl_input)
 		frame_output = system.reply()
+		
 		a = str(frame_output['system_action'][0])+"("
 		for i in frame_output['slot'].keys():
 			a.join(str(i)+',')
 		a.join(")")
+
 		turn_by_turn(currturn+1,a,'system')
 		currturn += 2
 		if currturn>= maxturn :
@@ -59,10 +61,10 @@ def run_an_episode():
 def turn_by_turn(currturn,turn,who):
 	print("turn{0} {1}:".format(currturn,who),turn)
 if __name__ == '__main__':
-	count = 1
-	succ = 0
+	count = 1.
+	succ = 0.
 	for _ in range(30):
 		r,s = run_an_episode()
 		if s is True: succ+=1
-		print (r,'rate:\t',float(succ/count))
+		print (r,'rate:\t',succ/count)
 		count+=1
